+++ src/item/popup/ApiBody/Corrupted.tsx
@@ -1,0 +1,13 @@
+import * as React from 'react';
+
+export interface Props {}
+
+export default class Stats extends React.PureComponent<Props> {
+  render() {
+    return (
+      <div className="unmet">
+        <span className="lc">Corrupted</span>
+      </div>
+    );
+  }
+}